opencl copy Y to pyramid

+++ Deshaker/CudaWriter.cpp
@@ -6,7 +6,7 @@ void CudaFFmpegWriter::open(EncodingOption videoCodec) {
     int result;
 
     //select codec
-    int deviceIndex = mData.deviceList[mData.deviceSelected]->type == DeviceType::CUDA ? mData.deviceList[mData.deviceSelected]->targetIndex : 0;
+    size_t deviceIndex = mData.deviceList[mData.deviceSelected]->type == DeviceType::CUDA ? mData.deviceList[mData.deviceSelected]->targetIndex : 0;
     const DeviceInfoCuda& dic = mData.deviceListCuda[deviceIndex];
     if (videoCodec.codec == Codec::AUTO) videoCodec.codec = dic.encodingOptions[0].codec;
     GUID guid = guidMap[videoCodec.codec];
@@ -15,7 +15,7 @@ void CudaFFmpegWriter::open(EncodingOption videoCodec) {
     FFmpegFormatWriter::open(videoCodec);
 
     //setup nvenc class
-    nvenc.createEncoder(mData.inputCtx.fpsNum, mData.inputCtx.fpsDen, GOP_SIZE, mData.crf, guid, deviceIndex);
+    nvenc.createEncoder(mData.inputCtx.fpsNum, mData.inputCtx.fpsDen, GOP_SIZE, mData.crf, guid, (int) deviceIndex);
 
     //setup codec parameters for ffmpeg format output
     AVCodecParameters* params = videoStream->codecpar;
